requirements.txt
Flask==2.3.3
anthropic==0.67.0
PyPDF2==3.0.1
python-dotenv==1.0.0